requirements.txt
pandas
streamlit
chromadb
google-generativeai
